Bug Fix:  Malicious clients can cause an unhandled exception
Issue:  In smtp-connection's auth handler, it directly uses network input as a key with an object.  If that key does not exist, then undefined will be returned, and immediately bind will be called, resulting in an exception.  Can reproduce by:
EHLO localhost
AUTH evil

Causes sasl["sasl_evil"].bind(this) to be executed.

Fix:  Check to see if the key exists; only call bind(this) if a truthy object is returned.