Added new condition to detect if object is null
A null object would still pass the typeof condition and throw. By adding
a new condition we make sure that _dataStream is not set to `null`.
Currently _dataStream is only set to boolean (default: false), Stream
(default: PassThrough) or Null but if another type is assigned to this
property of context, this code should be updated.